move tracer to src/
