+++ src/tracer.ts
@@ -1,7 +1,7 @@
 import * as grpc from "@grpc/grpc-js";
 import { context } from "@opentelemetry/api";
 import { AsyncHooksContextManager } from "@opentelemetry/context-async-hooks";
-import { OTLPTraceExporter } from "@opentelemetry/exporter-trace-otlp-grpc";
+import { OTLPTraceExporter as GrpcOTLPTraceExporter } from "@opentelemetry/exporter-trace-otlp-grpc";
 import { OTLPTraceExporter as ProtoOTLPTraceExporter } from "@opentelemetry/exporter-trace-otlp-proto";
 import { Resource } from "@opentelemetry/resources";
 import {
@@ -55,7 +55,7 @@ function createTracerProvider(endpoint: string, headers: string, attributes: Att
         headers: stringToHeaders(headers),
       });
     } else {
-      exporter = new OTLPTraceExporter({
+      exporter = new GrpcOTLPTraceExporter({
         url: endpoint,
         credentials: grpc.credentials.createSsl(),
         metadata: grpc.Metadata.fromHttp2Headers(stringToHeaders(headers)),
